rename read-file-encoding
