+++ test/specs/repo/extensions/scribunto.ts
@@ -1,5 +1,5 @@
 import { readFile } from 'fs/promises';
-import { utf8 } from '../../../helpers/readFileEncoding.js';
+import { utf8 } from '../../../helpers/read-file-encoding.js';
 
 describe( 'Scribunto', function () {
 	beforeEach( async function () {
