+++ test/specs/repo_client/interwiki-links.ts
@@ -1,5 +1,5 @@
 import { readFile } from 'fs/promises';
-import { utf8 } from '../../helpers/readFileEncoding.js';
+import { utf8 } from '../../helpers/read-file-encoding.js';
 
 describe( 'Interwiki links', function () {
 	it( 'Should be able to insert interwiki links', async function () {
